+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add some basic documentation on the following topics:
   - Accessing services managed by the injector
+* Enhance the error message on unresolved injectors to suggest using the `sting.debug` annotation processor option.
 
 ### [v0.07](https://github.com/sting-ioc/sting/tree/v0.07) (2020-03-12) · [Full Changelog](https://github.com/sting-ioc/sting/compare/v0.06...v0.07)